// fxpro-invest/backend/src/permissions/permission.enum.ts
export enum Permission {
  USER_DASHBOARD_VIEW = 'user.dashboard.view',
  PROFILE_MANAGE = 'profile.manage',
  DEPOSIT_CREATE = 'deposit.create',
  DEPOSIT_VIEW_OWN = 'deposit.view.own',
  DEPOSIT_VIEW_ALL = 'deposit.view.all',
  DEPOSIT_APPROVE = 'deposit.approve',
  DEPOSIT_REJECT = 'deposit.reject',
  INVESTMENT_VIEW_OWN = 'investment.view.own',
  INVESTMENT_VIEW_ALL = 'investment.view.all',
  INVESTMENT_PAUSE = 'investment.pause',
  INVESTMENT_CANCEL = 'investment.cancel',
  PROFIT_VIEW_OWN = 'profit.view.own',
  PROFIT_VIEW_ALL = 'profit.view.all',
  PROFIT_PROCESS = 'profit.process',
  WITHDRAWAL_CREATE = 'withdrawal.create',
  WITHDRAWAL_VIEW_OWN = 'withdrawal.view.own',
  WITHDRAWAL_VIEW_ALL = 'withdrawal.view.all',
  WITHDRAWAL_APPROVE = 'withdrawal.approve',
  WITHDRAWAL_PROCESS = 'withdrawal.process',
  WITHDRAWAL_PAY = 'withdrawal.pay',
  WITHDRAWAL_REJECT = 'withdrawal.reject',
  REFERRAL_VIEW_OWN = 'referral.view.own',
  REFERRAL_VIEW_ALL = 'referral.view.all',
  REFERRAL_REVERSE = 'referral.reverse',
  TRANSACTION_VIEW_OWN = 'transaction.view.own',
  TRANSACTION_VIEW_ALL = 'transaction.view.all',
  USER_VIEW = 'user.view',
  USER_SUSPEND = 'user.suspend',
  USER_ACTIVATE = 'user.activate',
  ADMIN_MANAGE = 'admin.manage',
  SETTINGS_MANAGE = 'settings.manage',
  REPORTS_VIEW = 'reports.view',
  AUDIT_LOG_VIEW = 'audit-log.view',
  BALANCE_ADJUST = 'balance.adjust',
}